Clean up the config file with a better interface to backtrace.
Now users can simply set -DSANITIZER_CONFIG_BACKTRACE_ENABLE=0 to disable it.
If enabling backtrace on a platform that isn't opted-in by default, users will
also have to set reasonable values for *_RUNTIME_DEFAULT_STATE, *_START_DEPTH,
and *_BACKTRACE_DEPTH.

Additionally, on x86/x64 floatmax_t has been redefined to just be a double
because frankly long floats shouldn't exist and I don't know what I was on when
I originally wrote the definition.

+++ include/sanitizer/platform.h
@@ -11,7 +11,10 @@ typedef sys_uptr sys_uhwptr;
 typedef sys_uptr ValuePtr;
 
 #if defined(__x86_64__) || defined(__x86_32__) || defined(__i386)
-typedef long double floatmax_t;
+
+// Long doubles shouldn't exist, so let's pretend they don't
+typedef double floatmax_t;
+
 #define SIntFormat "%li"
 #define UIntFormat "%lu"
 #define FloatFormat "%lE"
